Lil' refactoring
